+++ middleware/documentation/routes/home-element/examples/template.js
@@ -1,0 +1,8 @@
+(context, html) => html`
+  <div>
+    <h3>Middleware output</h3>
+    <ul>
+      ${context.outputs.map((output) => html`<li>${output}</li>`)}
+    </ul>
+  </div>
+`;